refactor(developer): rename validate-layout-file.ts
Rename `ValidateLayoutFile()` to `CompileLayoutFile()` to better reflect
the transform steps it performs, add documentation, and address review
comments.
